refactor(plan_SERVERSPLIT slice 2): the tabs bodies leave server.py
Five more bodies move into embedded/tool_sections/tabs.py. This section is the
first to use _require_browser and _require_landing_ok, and the first to read a
tuned knob DIRECTLY rather than through _with_cdp_timeout's default: new_tab
hands CDP_OPERATION_TIMEOUT to _require_landing_ok. That read is the point of the
slice. It resolves as rt.CDP_OPERATION_TIMEOUT against tool_runtime at CALL time,
so tests/conftest.py's patched_server(CDP_OPERATION_TIMEOUT=...) still reaches a
body that no longer lives in server.py — the property a `from ... import` binding
would have silently destroyed.

A pure move: bodies verbatim, docstrings and signatures byte-identical,
tests/goldens/tool_surface.json unchanged. server.py loses the _require_browser
import with them, tabs being its last consumer.

Verified: 2334 passed / 1 skipped on the unit lane; tool surface IDENTICAL to the
golden; SECTION_TOOLS 94 with per-app [94, 94, 94]; --list-sections subprocess
"Total: 94 tools"; gen_release_contract --check, ruff format/check, vulture, ty
and the suppression-owner gate all clean. Two real-Chrome real-stdio lanes, run
one at a time: test_e2e_transport.py (the canonical journey covers list_tabs and
get_active_tab) and test_soak_stability.py, whose soak cycle is the only real
transport coverage of the new_tab -> switch_tab -> close_tab -> list_tabs churn,
including the post-close_tab re-read that the nodriver rediscovered-target quirk
lives in. Both green.

server.py: 3248 -> 3144 LOC, cap ratcheted DOWN to the measured actual;
tests/source_scan.py's floor ratchets UP 3 -> 4.
